Apply dhewm3 / ce976c470197176f8892a837aaf34d2e6ddae665: Fix script interpreter stack for x86_64
Alot of stack and event variables are pointers. Align the size of
all script and event variable types to sizeof(intptr_t) so that
the CPU needs only one fetch insn on 64bit archs.

Tested on x86 and x86_64 and found no different script behaviours
compared to the binary from id.
Savegames on x86_64 do work, but are not compatible to x86 and vice
versa (among other issues, the stack is written to file as-is).
x86 builds can still load savegames from the official binary and
vice versa.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6200 49c82d7f-2e2a-0410-a16f-ae8f201b507f