Apply dhewm3 / c7956a1cb51d9febaee47467374b9971e72fb304: Fix freeze when quitting the game during timedemo
When one tries to quit the game (via quit in console) while a timedemo is running, the game freezes indefinitely while displaying the timedemo results.
This happens because the MessageBox used for that waits for a an event normally triggered by a timer - but it seems like that does not happen anymore during Shutdown(). The fix makes sure that this message box isn't displayed.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6306 49c82d7f-2e2a-0410-a16f-ae8f201b507f