Added FFmpeg source and static library for Linux 32-bit.
Note: I took version 2.5 of FFmpeg, which is only approximately equivalent to the version greebo used.
I think he added version between 2.5. and 2.5.1 initially.
This means that linux static libs are a bit out of sync with headers and windows static libs.

The added static libs were configured as:
  ../../lib/ffmpeg/configure --disable-debug --enable-small --arch=x86_32 --extra-cflags="-m32" --extra-cxxflags="-m32" --extra-ldflags="-m32"
So they are stripped of debug information and optimized for size.

Anyway, there is no point in fixing these issues now.
I think I should simply do a full update of FFmpeg when this Linux build mess is over.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/linuxbuild_2017_06@6954 49c82d7f-2e2a-0410-a16f-ae8f201b507f