Improving rsqrt/sqrt methods of idMath (following commit 6842).
Previously I did not think that sometimes negative numbers get into sqrt/rsqrt.
And actually they get, and before this commit it resulted in Inf or NaN result (which is very BAD).

Now all the functions have simple behavior in presence of SSE2 (which is enabled everywhere now):
  RSQRT(x) := 1 / sqrt(max(x, eps))     eps = minimal positive float32  (normal, not denormalized)
   SQRT(x) := sqrt(max(x, 0))
So sqrt is perfectly stable around 0, and returns 0 for negatives.
rsqrt returns big number (about 1e+18) for zero, and the same number for negatives.

Hopefully, this would fix most of the NaNs appearing in game.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7103 49c82d7f-2e2a-0410-a16f-ae8f201b507f