Reverted DevIL from 1.8.0 (2017, C++) to 1.7.8 (2010, C).
The reason is having less artefacts in SVN.
A pure C library can be compiled in release mode and used both in debug and release.
A C++ library must be compiled separately for debug and release (depending on iterator-debugging, runtime, etc).
I guess in future it would be easier to remove DevIL completely, given how little it is used.



git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/3rdparty_4822@8131 49c82d7f-2e2a-0410-a16f-ae8f201b507f