Added all recently built dependencies (Both Windows and Linux, both 32-bit an 64-bit).
I have built TheDarkMod on Windows in Debug, Debug with Inlines, Release on 32-bit and 64-bit architecture.
Built it on Linux in debug and release on 32-bit and 64-bit.
Also tested some of the builds on Windows and Linux --- it works.

Also built debug/release tdm_update solution on Windows, and release tdm_update stuff on Linux (both 32-bit and 64-bit).
I tested packager and updater some time ago.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/3rdparty_4822@8147 49c82d7f-2e2a-0410-a16f-ae8f201b507f