#5027. Implemented idSIMD_SSE2::CreateVertexProgramShadowCache.
It took significant time in "Behind Closed Doors" scene in "Debug with Inlines".
So I decided to implement it in idSIMD_SSE2, since optimization is force-enabled here for this build.

Also minor change in idDenseHash::Find, which might make a difference in non-optimized build.
But I'm afraid it does not =(


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8179 49c82d7f-2e2a-0410-a16f-ae8f201b507f