Deleted base64 code.
The only place where it is used in idCommonLocal::WriteConfigToFile.
It is called to convert obfuscated si_version from binary to text and save to config file.
This is done only under ID_WRITE_VERSION, which is never defined (perhaps it was some server code).

The rationale: we don't want to drag unnecessary piece of code under yet another license, even if ID's employer was its author =)
Right now this piece of code even requires to be mentioned in binary-form distribution.
If someone needs base64 again, he would easily find some public-domain code on the internet.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8199 49c82d7f-2e2a-0410-a16f-ae8f201b507f