Removed CRC32 code.
It was only used in idDict::Checksum method, which is not used anywhere.
Besides, we can use always crc32 implementation from zlib if we ever need to (we do so in tdm_update after boost removal).

The main reason for removing the code is that it is clearly marked "Copyright by Mark Adler", but has absolutely NO license.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8202 49c82d7f-2e2a-0410-a16f-ae8f201b507f