#4957. Reduce chance that close vertices would be snapped to different points in T-junctions removal.
Note the T-junctions removal algorithm also serves as a generic mesh cleaner.
It snaps triangle vertices exactly to each other, so that further mesh processing could work reliably.

The problem is that when vertex is snapped to some grid point, its coordinate is rounded to the nearest of two snap points.
If model has vertices exactly at the middle between snap points, then they get snapped into different directions depending on roundoff error.
Having vertex with coordinate (2*integer+1)/64 is quite likely in a real model.
The fix is to shift coordinate slightly, so that it gets rounded down in such cases.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8558 49c82d7f-2e2a-0410-a16f-ae8f201b507f