#5296. Avoid NaN when solving quadratic equation in collision detection.
This happens in functions:
  idCollisionModelManagerLocal::RotateEdgeThroughEdge
  idCollisionModelManagerLocal::EdgeFurthestFromEdge
  idCollisionModelManagerLocal::RotatePointThroughPlane
  idCollisionModelManagerLocal::PointFurthestFromPlane
All of them contained the same copy-pasted code for solving quadratic equations.
This code is extracted to SolveQuadraticEquation local function.

The function is also fixed to avoid creating NaN as second solution when the first one is zero.
Note: you can toggle the #if in this function to restore old behavior.

This situation is easily reproduced on Inn Business FM.
Execute: setviewpos 594.12 -87.79 68.25   9.1 -178.1 0.0
Now take some bottle and move it around, colliding with other objects.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8858 49c82d7f-2e2a-0410-a16f-ae8f201b507f