#5296. Avoid NaN in CBinaryFrobMover::GetFractionalPosition during double-door post-spawning.
Unfortunately, the spawnargs which define door movement are read during post-spawn.
When a double door is present on a map, both of its halves are independent entities, and they post-spawn one-by-one.
When first half is post-spawned, it calls GetFractionalPosition of the second half, which is not post-spawned yet.
Hence all movements are zero, and GetFractionalPosition returns NaN.
Note that after that second half is post-spawned, and it overrides all the bad effects which the NaN caused.

Happens on Behind Closed Doors FM during game start.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8861 49c82d7f-2e2a-0410-a16f-ae8f201b507f