#5433,#5170: Reverted commit 8947.
I'm going to fix dangling pointer issue properly in the following commits.
So this hack for dead windows is no longer necessary.

--------------------------------------------------------

Revision: 8947
Author: stgatilov
Date: 23 august 2020 г. 16:17:45
Message:
#5170. Fix for crash with "decl_stack 1" on loading of "The Rat's Triumphant".
The reason was dangling pointers to idWindow.

Unlike all the other objects, idWindow-s seem to be recreated and deleted too eagerly.
For now I have "fixed" it with a very dirty hack.
When LoadStack needs to print window level, it asks global uiManager if this window pointer is currently present in all of its GUIs.
If it is not, then name is not queried.

It means that sometimes the name of the window will be missing in the load stack.
There is also small probability that window name will be reported incorrectly, if the same memory buffer will be reused for another window.
But, most importantly, it does not crash now.
----
Modified : /trunk/framework/LoadStack.cpp
Modified : /trunk/ui/UserInterface.cpp
Modified : /trunk/ui/UserInterface.h
Modified : /trunk/ui/UserInterfaceLocal.h



git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9065 49c82d7f-2e2a-0410-a16f-ae8f201b507f