#4550. Do not use/include sys_intrinsics.h.
This header was recently copied from BFG, and it contains a few bits of SSE code.
The problem is that unlike all other cases of SSE usage, this header does not provide generic (non-SSE) implementation.
It simply includes <xmmintrin.h>, considering that it must exist everywhere.

This prevents from compiling TDM on platforms like ARM or PowerPC.
Of course, such platforms are not supports right now, but such things change sometimes.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9275 49c82d7f-2e2a-0410-a16f-ae8f201b507f