#4550. Protect CPUID instruction and xmmintrin.h include from non-x86 architectures.
CPUID function now returns all zeros if target architecture is not x86.
Also, HasCPUID returns false on such architectures, so normally CPUID should never be called then.

The purpose of this change is to fix compilation on platforms like ARM or PowerPC.
Of course, we do NOT support such architectures.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9279 49c82d7f-2e2a-0410-a16f-ae8f201b507f