Added /WHOLEARCHIVE:idlib to VC props of TheDarkMod project.
This helps for tests defined in idlib project.
Idlib is a static library, containing several .obj files.
Each .obj file is only linked into executable if one of its symbols is used.
As the result, doctest tests are not registered in some cases:
  https://github.com/onqtam/doctest/blob/master/doc/markdown/faq.md#why-are-my-tests-in-a-static-library-not-getting-registered

Note that CMake build does not separate code into idlib and TheDarkMod, all files are in one project.
So this problem does not apply there.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9313 49c82d7f-2e2a-0410-a16f-ae8f201b507f