#5600. Removed system of "noModified" callbacks on cvars.
This probably sounded like a good idea, but it is not.
There is already IsModified/ClearModified mechanism, which is much more reliable because it ensures deterministic moment when callback is performed.
The option of firing callback immediately after change is dangerous: it would lead to serious problems if it were widely used.
This is especially bad with multithreading, because you cannot predict which thread would call your callback.

Anyway, the system was only used for dnPostProcessManager::OnImageAnisotropyChanged depending on anisotropy cvar.
The callback is now emply, since old postprocessing/bloom is long gone.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9349 49c82d7f-2e2a-0410-a16f-ae8f201b507f