#5323. Refactored various background stuff into the system of background layers.
Now every menu state has "backgrounds" macro attached.
Macros look like MM_BACKGROUNDS_SUCCESS and must be defined in GUI scripts.
The right side of the macro is the comma-separated list of background layers to enable for this state.

The C++ code looks which background layers are different between previous and current state.
The removed layers are disabled, the added layers are enabled.
The comma-separated list of currently enabled backgrounds is stored in "gui::backgrounds" variable.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/5323_mainmenu@9420 49c82d7f-2e2a-0410-a16f-ae8f201b507f