#5323. Implmented "meta" backgrounds defines for.
If the value of backgrounds macro starts with MM_BACKGROUNDS_, then we treat it as another macro.
So we get its value, and reconsider it again --- until we arrive at something not starting with MM_BACKGROUNDS_.

In fact, this allows chaining macro values --- a bit of expansion over the ordinary "only support plain value" scheme.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/5323_mainmenu@9433 49c82d7f-2e2a-0410-a16f-ae8f201b507f