#5665. Remove support for PCX image format.
This format was used in ZSoft Paintbrush in MS-DOS times. It is not popular today at all.
There is no .pcx file in assets SVN, and no .pcx file in any of the released FMs.
The .bmp and .tga formats do the same (.tga can do RLE encoding just like .pcx).


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9453 49c82d7f-2e2a-0410-a16f-ae8f201b507f