Followup to rev 9609. Do NOT even try to render without index cache in core profile, don't spam console warnings when this happens.
The reason to avoid warning spam is that the case of no index buffer legitimately happens when whole index buffer is resized.
This happens rarely and very briefly, but posting warnings at that moment result in serious lag, plus mappers won't be happy to scroll through tons of warnings.

The reason to avoid rendering from CPU index buffer is that there is no guarantee that driver will simply ignore such call.
It can as well crash or get into inconsistent state, so better be safe and not do anything.

Note: if someone enables deprecated profile, then renders without index buffer are done.
Let it live for some time for easier debugging... maybe we will completely remove this code in future when we are sure no old code remains.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9642 49c82d7f-2e2a-0410-a16f-ae8f201b507f