Fixed "clip model outside world bounds" issue broken by svn rev 9592.
According to idCollisionModelManagerLocal::CollisionModelForMapEntity, the collision model of worldspawn is named:
  1) by the "name" spawnarg of worldspawn entity (e.g. "brush", "worldspawn", "omg" --- whatever mappers sets).
  2) "worldMap" if worldspawn entity does not have "name" spawnarg.
However, when idClip determines bounding box of the whole world, it always searches for "worldMap" collision model.
As the result, the current code works properly if there is no "name" spawnarg, but breaks when the spawnarg is set.

Now the interesting part: why did it work properly before svn rev 9592?
The reason is that previously some of collision model methods returned 0 if specified model was not found.
Since 0 is handle of existing collision model, I changed that to -1 in svn 9592, so that we are always sure it cmhandle is dead or not.
However, the very first collision model in .cm file is always the world model, that's why it worked before when "0" was returned in case model was not found.

So, this fix just says: always take the very first collision model as world, regardless of its name.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9643 49c82d7f-2e2a-0410-a16f-ae8f201b507f