Fixed bug which caused tdm_installer to reanalyze installation after TDM run.
The analysis was run because TDM repacked tdm_base01.pk4.
The reason it repacked tdm_base01.pk4 is: it thought "strings/all.lang" had "png" extension =)
It turns out that idStr::CheckExtension assumes that extension starts with dot, while in this case it was not.
It does not compare the first character at all =(

Here is the list of changes:
1) idStr::CheckExtension now asserts that first character of extension is "."
2) Function DoNotCompressFile now prepends "." to extension name before calling CheckExtension (that's the main bugfix).
3) stdext::iends_with now calls new idStr::IendsWith instead of idStr::CheckExtension.

As far as I see, DoNotCompressFile and stdext::iends_with were the only calls to CheckExtension without dot.
Also stdext::iends_with is only called in MissionManager for l10n packs, so even if I break it, it won't be a big problem =)


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9817 49c82d7f-2e2a-0410-a16f-ae8f201b507f