#5599. Simplified rotation logic for new grabber.
Basiically, ignore object mass.
Just rotate the thing almost as fast as we can with sa bit of lag.

This is goind to reduce this problem:
  https://forums.thedarkmod.com/index.php?/topic/21198-beta-testing-210/&do=findComment&comment=469415
Although it does not do anything to fix it.
The core problem is that to-be-applied rotation is stored as target orientation instead of desired rotation angles.
If "to-be-applied" rotation exceeds 180 degrees, then grabber starts going backwards since that's the shortest route.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9818 49c82d7f-2e2a-0410-a16f-ae8f201b507f