#5881. BLAKE2 recipe converted to versionless format, use 20190724 tag version instead of git master.
Versionless means that now source code url for any version is stored in conandata.yml.

The reason to not use git master is that builds are far less reproducible (if some major change lands into master, which is unlikely though).
In fact, this is a version downgrade, but all the changes in master since the latest tag are warning fixes.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9868 49c82d7f-2e2a-0410-a16f-ae8f201b507f