#5819. Changed screenshot filename format, make sure coords are exactly as in getviewpos.
Extracted the code from getviewpos command into idGameLocal method.
To be honest, I'm not sure how and when it is different from simply taking firstPersonViewOrigin member...
One sure difference is the case of third-person view, which is mostly dead anyway.

Also changed format of filename to make parts more visible:
1) hours/minutes/seconds are not separated by hyphen (-)
2) timestamp and position are surrounded in parentheses and separated by spaces
Yeah, parens are not very portable, but I think we need to try.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9927 49c82d7f-2e2a-0410-a16f-ae8f201b507f