#5849. Warn about custom window registers without float keyword, and about register with EOF instead of value.
Had to refactor idWindow::ParseRegEntry to make it more clear.

For example, our compass had the only instance of this error:
  noshadows 1
This is custom variable, which is read deeply inside idRenderWindow code, scripting language does not know anything about it.
This can easily be an error, so now we print warning in this case.

In order to avoid warning, use float keyword:
  float noshadows 1
Note that there is also definefloat keyword, which does the same (no idea about any difference).
And there is definevec4 keyword for vector register.
If we need more types, we can add similar defineint/definebool/whatever in future.

However, I don't think custom window registers are useful.
This particular case is not convincing, since this variable is hardcoded into C++ engine anyway.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10010 49c82d7f-2e2a-0410-a16f-ae8f201b507f