#5869. Better checking of values for internal vars, warn if resetCinematics has arguments.
For some reason, resetCinematics was declared to have 0 to 2 arguments, while it never uses any at runtime.
Now it is declared to accept [0..0] arguments, so this code will produce a warning:
  resetCinematics 0;
  resetCinematics MenuStartWindow;
  resetCinematics     resetTime 0;     //imagine this is two lines with missing semicolon

To be honest, internal vars are not used too often.
An warning would be generated e.g. in this case:
  noclip 7     //should be 0 or 1


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10013 49c82d7f-2e2a-0410-a16f-ae8f201b507f