#6149. Fixed glitch on game load.
It was caused by two logical frames being concatenated together into one list of backend commands.
That caused wrong FBO state on the second half.

After this commit, each of these frames is rendered as its own backend commandlist.
That's enough to fix the issue for me, SMP both on and off.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10182 49c82d7f-2e2a-0410-a16f-ae8f201b507f