#6088. Override lightgem return value for 6 frames after game start or load.
When game is loaded, we restore old values of m_LightgemShotValue, and set override to their maximum value for the next 6 frames.
Similarly, when game is started/restarted, set override for 0.0 (full dark) for the first 6 frames.

Note: it is important that lightgem continues working and pushing the data through all of its buffers while override is active.
Only the very final return value is replaced.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10224 49c82d7f-2e2a-0410-a16f-ae8f201b507f