#6225. Assert that we don't leak clipmodel.
Note that it clipmodel leaks, it remains in idClip, so when some query hits it, the code tries to access potentially dead owner entity.
Also removed NULL assignment, since if old value was not NULL, then it is surely a leak.

These asserts fails pretty often during melee combat, which explains why sometimes it crashes.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10243 49c82d7f-2e2a-0410-a16f-ae8f201b507f