Added idSoundSample::DurationIn44kHzSamples in addition to LengthIn44kHzSamples, fixed some usages.
Let's suppose we have a sound which plays for T seconds, and includes C channels:
  DurationIn44kHzSamples  ==>  44.1K * T
  LengthIn44kHzSamples    ==>  44.1K * T * C
So "length" is doubled for stereo sounds, compared to "duration", which corresponds to astronomic time.
In fact, "length" is more about amount of data, while "duration" is more about actual time.

Also fixed usages in:
  idSoundShader::List --- I guess used in some listing command for sound shaders...
  ListSounds_f --- used by listSounds command
  idSoundWorldLocal::AddChannelContribution --- getting DB adjustment for leadin / external params (not sure about effect)

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10279 49c82d7f-2e2a-0410-a16f-ae8f201b507f