#6271. Deleted INATTR_POSITION macro in shaders.
It was introduced when we were supporting both core and compatibility profile.
At that moment we used fixed-function transform on compatibility, and our custom function on core.
They were incompatible by precision, so to avoid the switch between two cases everywhere, we introduced the macro.

We won't use fixed-function stuff anymore =)

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10297 49c82d7f-2e2a-0410-a16f-ae8f201b507f