#6280. Added "Sanitize" configuration with Address Sanitizer turned on.
Debug Fast is used as a basis for it: it is as slow as release, but improves debugging experience.

Also note that I defined _DISABLE_VECTOR_ANNOTATION to avoid checks for std::vector elements access in [size()..capacity()) range.
This check requires all third-party dependencies to be rebuilt with ASAN, which is a nuisance, and we should not use much STL anyway.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10352 49c82d7f-2e2a-0410-a16f-ae8f201b507f