Added root directory to include paths, removed "framework" from include paths.
The reason for having root in include dirs is allowing to include our headers by their path from root.
With subdirectories getting deeper, writing "../../renderer/backend/stages/BloomStage.h" is not fun.

The reason for not having "framework" in include paths... just why? =)
"idlib" is there because it was a separate project some time ago.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10377 49c82d7f-2e2a-0410-a16f-ae8f201b507f