#6296. Changed data structures for area-entity and area-light connections for faster iteration.
Before this change, every area-entity instance was represented by a separate areaReference_t struct of size = 48 bytes.
In order to iterate over all incident area-entity edges for an area or an entity, these structs were linked in two linked lists (doubly-linked in area, singly-linked in entity).
As the result, iterating over all entities in area in CreateLightDefInteractions required traversing linked list, with nodes laid in memory in pretty random order.

The core of the new data structure is:
  idList<int> portalArea_t::entityRefs;
which stores indices of entities in idRenderWorldLocal::entityDefs array.

This allows to iterate over all entities in area without random access and touching excessive memory.
In fact, this kind of iteration is as fast as it can possibly be.
It accelerates CreateLightDefInteractions on Iris from about 7% to about 5% for me, and further improvements will follow.

Note that using indices instead of pointers also allows to store some important attributes of entities in separate arrays in the world.
One of the problems is that idRenderEntityLocal is huge, and thus touching any member of it in CreateLightDefInteractions is very expensive.
But if an attribute is stored in separate array, it can be done in much more cache-efficient manner (this is basically a partial structure-of-arrays approach).

The rest of the data structure is to support iteration of all areas of an entity in O(1) time per area, and to support adding/removing references in O(1) time.
I don't care much about how fast the constant is here, so there are still linked lists here --- that allows to avoid having many small arrays (i.e. one per entity).

One sad thing is that since indices are stored instead of pointers, it is a bit harder to traverse the structures in debugger.
Still, I managed to support area <-> entities,lights virtual containers in natvis, although I had to use global render world for that.
I hope we'll never need to have several render worlds, otherwise this will bite us =)


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10384 49c82d7f-2e2a-0410-a16f-ae8f201b507f