#6296. When considering a noshadows light, skip areas which are not in view.
We already skip entities which are not in the view in this case.
If an entity is in the view, then at least one of the areas it belongs to must be in view too.
This idea might be a bit questionable, so keep an eye on new problems to appear =)

It allows to dramatically reduce time in CreateLightDefInteractions to 3.5% on Iris.
That's probably because one light is always ambientWorld, which is noshadows and covers everything (other lights like this exist on some maps).


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10386 49c82d7f-2e2a-0410-a16f-ae8f201b507f