Added debug commands tdm_open_doors and tdm_close_doors.
Of course, it can break missions by opening things which can only be logically opened via script or special handle, etc.
So it is only useful for debugging/profiling.
For instance, it is useful to check performance with all portals open: it is known to be much worse than with portals closed.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10395 49c82d7f-2e2a-0410-a16f-ae8f201b507f