#5172. Fixed bug with noshadows lights using light portal flow.
If a light has material which can cast shadows but is marked as "noshadows" in the map,
then it used light portal flow and only lights areas + places directly visible from light origin.
This is wrong.

Now we check for noshadows both in material and in light entity.
As far as I know, if both flags mean shadows enabled, then world areas surely do cast shadows --- that's what we need here.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10414 49c82d7f-2e2a-0410-a16f-ae8f201b507f