#6164. Merged definefloat and float, added vec4 keyword.
Note that to restore uniformity I do breaking change to definevec4 (which is not used in our core GUIs at least).
Previously the register was added to top-level windows (Desktop of current UI), instead of the current window.
Not it is added to the current window, just like it works for float already.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10461 49c82d7f-2e2a-0410-a16f-ae8f201b507f