Fixed unused result and pugixml deprecation GCC warnings.
The result of system in Sys_DoStartProcess is still unused.
I'm not sure there is anywhere to put the return code, or anything to do on fail.

Result of fread in BackgroundDownloadThread is also unused: also not sure where to put it.
Anyway, this background download code from Doom 3 is not used: TDM has its own background download code.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10570 49c82d7f-2e2a-0410-a16f-ae8f201b507f