Fixed confusion between enums constraintType_t and declAFConstraintType_t.
These two types were compared in idAF::Load, hoping that their common values mean the same.
It seems that some values did not match.

Luckily, this part of code is only called in some editors.
Because in-game Load is called during spawning, where AF is empty, and this code runs only if AF was previously filled.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10573 49c82d7f-2e2a-0410-a16f-ae8f201b507f