Fixed player getting stuck at zero velocity sometimes.
In particular, it happens at the beginning of One Step Too Far:
  https://forums.thedarkmod.com/index.php?/topic/16457-fan-mission-one-step-too-far-by-dragofer-20140804/&do=findComment&comment=491828

Note that before svn rev 10498, vertical component of velocity was doubled instead of removed in line 664.
As the result, this "if" could only trigger when velocity is less than 1 horizontally and less than 0.5 vertically.

To be honest, I don't see ther point in this code anyway.
And multiplying velocity by 0.995 every frame is just invalid, since we have uncapped FPS today.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10636 49c82d7f-2e2a-0410-a16f-ae8f201b507f