#6425. Fix forced-crouch mantle animation (again)
Sometimes the player view would clip into the ceiling, exposing the
skybox during a forced-crouch mantle into a duct.

This employs a new strategy of matching the crouch animation to the
pull mantle animation. The old strategy assumed the mantle animation
had a defined beginning and end. However, during a mantle into a duct,
sometimes the vent door and the duct would cause rapid switching
between mantle types. The new strategy makes increments based on
previous and current origins. It's the journey, not the destination.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10649 49c82d7f-2e2a-0410-a16f-ae8f201b507f