#6606. Deleted concept of "opaque" GUI overlay.
Basically, this is a flag which allows to optimize rendering GUI overlays if there are many opaque ones (only last one is rendered).
We don't have any opaque layers at all, there is little point in having them, and GUI is very fast anyway.
So there is no good in having code for an optimization which was unnecessary for 15 years =)


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10959 49c82d7f-2e2a-0410-a16f-ae8f201b507f