Disabled all timestamp / CRC checks in AAS/CM build.
The reason for that is: these checks are unreliable.
Most importantly, AAS might change due to material contents change, or due to engine change.
These "can we save old file?" allowed false positives, which results in a lot of confusion sometimes.
And the advice to manually delete all the files before dmap is a bit weird.

Discussion here:
  https://forums.thedarkmod.com/index.php?/topic/22944-dmap-and-overwriting-files


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11000 49c82d7f-2e2a-0410-a16f-ae8f201b507f