#6546. Dropped hacks to keep bodies and dropped melee weapons always tracked in LES.
The main approach is: the first query gets zero light value is 0, but starts tracking for the next 10 seconds.
It is expected that repeated queries will come during these 10 seconds, then next queries will get proper values.
Right now most visual stims have period <= 2 seconds (and I'll probably reduce this), so it should be fine.
The only downside is that the reaction of AI to the object is delayed by one query period (which is important for visual stims).
So it's better to avoid hacks and try to do everything based on this approach.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11001 49c82d7f-2e2a-0410-a16f-ae8f201b507f