#6546,#6623. Properly supported blood markers in Light Estimate System.
To achieve this, I had to support a new feature: overriding samples for LES explicitly for an entity.
Such an override is set by calling LightEstimateSystem::SetExplicitSamplingForEntity.
It becomes the part of game state, and is save/restored in savegame along with an entity (the code is in LightEstimateSystem::[Save|Restore]ExplicitSampling).
The only use case right now is in BloodMArker.cpp: we simply generate a bunch of points across decal surface and set them as override.

The internal implementation is a bit more complicated though...

SetExplicitSamplingForEntity does something similar to what FindOrAddModel does, so I extracted their common ending into FinishModelSampling.
It replaced the data in entity, and marks internal "needsUpdate" flag.
This flag is catched during global Think, and it causes complete readdition of the entity (like on model change).

The FindOrAddEntity method now checks if entity has explicit sampling, and uses it if it does.
In this we set modelIndex to special value of MODEL_INDEX_EXPLICIT, which now we have to check for explicitly everywhere =)
Also I added a new helper method GetSamplingOfTrackedEntity, which handles both cases --- it is used in StartNewQueries and DebugVisualize for reading the sampling.

Finally, ShouldSampleBeExcluded now checks if surfaceIndex == -1.
This happens for explicit samples, since right now SetExplicitSamplingForEntity accepts only array of 3D points.
In principle, in the future we can allow to bind explicit samples to model's surfaces, so that their material flags affect whether the points are enabled or not.
But there is no point in doing it now.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11006 49c82d7f-2e2a-0410-a16f-ae8f201b507f