#5867. For two-sided material with interactions, make sure to set ID for duplicated surfaces.
In case ShouldCreateBackSides triggers, model surfaces are duplicated during R_CleanupTriangles.
However, ID is not set for these duplicates, so it can get any value.
Sometimes it gets negative value, which is treated as "overlay surface" and removed by frontend.
I'm pretty sure this is not the intended behavior.

Now it is initialized properly and should never get negative this way.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11023 49c82d7f-2e2a-0410-a16f-ae8f201b507f