#6647. Preload the loop body in addition to the first-played sound sample in StartSound.
When level load ends, all the sounds not referenced during the load are purged.
However, some sounds are played on demand only, most importantly all the main menu sounds.
To make them work, idSoundEmitterLocal::StartSound checks if the sound sample was purged, and loads it if it was.

Unfortunately, this on demand loading only covers "channel->leadin", which is the first sound sample to be played.
This is usually OK, except for the case of looping sound shader with leadin sample: in this case only leadin is loaded, and the loop body remains unloaded and thus does not play.
With this fix, the loading logic covers the loop body as well.

Note: I assume that loop body is entries[0] (the first sound file in the shader body).
It seems that multiple bodies with random selection are not expected to be used together with looping sounds...

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11032 49c82d7f-2e2a-0410-a16f-ae8f201b507f