Now "portal_dist" respects LOD multiplier.
It seems that the only viable usage for this spawnarg is performance optimization.
In which case it acts like hide_distance of objects.
Hence, it would be great it player could delay or disable this optimization if their machine is a beast.
That's why it makes sense to use the multiplier.

See also:
  https://forums.thedarkmod.com/index.php?/topic/23041-lod-func_portal-and-better-range/

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11033 49c82d7f-2e2a-0410-a16f-ae8f201b507f