#6648. Fixed starting of the next mission in campaign when BRIEFING_VIDEO state is disabled.
The old code just switched to BRIEFING_VIDEO unconditionally (since it is the first state in the mission start sequence).
But if this state does not exist, then it just stops and falls back to main menu.

The new code simply simulates forward transition from MAINMENU_NOTINGAME, so it is capable of doing proper skips, all the way up to starting game.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11037 49c82d7f-2e2a-0410-a16f-ae8f201b507f