#6648. Fixed starting new mission when SUCCESS state is disabled.
The code which does switch to the next mission happens on "onsuccessscreencontinueclicked" GUI command.
This command is executed only when user clicks button on the SUCCESS screen.

Now forward transition from SUCCESS screen goes into AFTER_SUCCESS auxiliary state.
It is handled in special way and we call GUI handler in the immediate future, which generates back "onsuccessscreencontinueclicked" GUI command.
We can't call this code directly, because we want the current GUI event to be processed completely first.

Note: requires changes in GUI code to work properly.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11038 49c82d7f-2e2a-0410-a16f-ae8f201b507f