#6608. Call R_IssueRenderCommands without SwapBuffers inside CaptureRenderToBuffer.
This RenderSystem call is used only for "last game frame for menu" and for savefile screenshots.
It finishes and executes all backend commands inside.
However, in this case we don't want to show this (potentially tweaked) frame to the player.
So we omit SwapBuffers and thus it is rendered to back frame but them overwritten on the next draw.

This removed weird frame on quicksave and menu open.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11052 49c82d7f-2e2a-0410-a16f-ae8f201b507f