#5485. Renamed viewDef_t: isMirror -> isMirrorInverted.
The reason for doing this is that similar flags (isSubview, isXray, isPortalSky) show whether this specific view was generated as subview of specific type.
But isMirror right now means whether the orientation of the view is inverted.
A mirror always inverts this flag, so inside two nested mirrors we should have isMirror = false.
This makes the flag very confusion because while it looks like it shows whether this is a mirror subview, it can't be used for that!


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11055 49c82d7f-2e2a-0410-a16f-ae8f201b507f