#5485. Supported mirrorResolutionFactor material flag to make low-resolution mirrored surface.
The effective resolution should be:
  RootViewResolution * min(mirrorResolutionFactor).
The r_fboResolution kick in on top of then in kinda "builtin" way.

RootViewResolution is the most-deep subview that is not "screen-space", which is usually main view or remote screen.
The minimum of factor is taken among all the subviews below that.
So if we have the same mirrors nested in each other, they'll all have the same resolution.

See also:
  https://forums.thedarkmod.com/index.php?/topic/23044-reduced-resolution-of-subviews/

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11057 49c82d7f-2e2a-0410-a16f-ae8f201b507f