#5485. Supported "remoteResolution" setting for remote subview materials.
It allows to specify resolution of remote screen in "pixels per doom unit".
Also it accepts maximum width/height in pixels to make sure that a huge screen won't bring the whole game down accidentally...

Aside from that, the resolution as an argument of remoteRenderMap is deprecated.
Now it is converted to true resolution during parsing to reproduce 2.13 behavior.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11058 49c82d7f-2e2a-0410-a16f-ae8f201b507f